backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pymysql==1.1.2
python-multipart==0.0.22
pillow==12.1.1
python-dotenv==1.2.2
requests==2.32.5
pydantic==2.12.5
pytest==9.0.2
mysql-connector-python==9.6.0